refactor: complete architecture cleanup and reorganize features
• Move modules/ HTML templates to src/features/ for better organization
• Remove generated bundle files from git tracking (now properly ignored)
• Clean up unused components and styles
• Update imports and references throughout codebase
• Improve .gitignore configuration for generated files
• Update CLAUDE.md with latest architecture documentation

This completes the architecture refactoring by:
- Consolidating all feature templates under src/features/
- Removing redundant/unused files
- Ensuring generated files are properly ignored
- Maintaining clean separation between development and runtime layers
