--- modules/get-started/components/futureFeatures.ts
@@ -1,65 +1,0 @@
-export const futureFeaturesComponent = `
-<!-- Future Features Section -->
-<div class="future-features">
-    <h2>🔮 即将推出的功能</h2>
-    <p style="text-align: center; color: var(--color-text-secondary); margin-bottom: var(--space-8); font-size: var(--font-size-lg);">这些功能正在开发中，敬请期待</p>
-    
-    <div class="feature-grid">
-        <div class="feature-card coming-soon">
-            <div class="coming-soon-badge">即将推出</div>
-            <h3>🛠️ 快速配置工具</h3>
-            <p>交互式配置向导，帮助您快速设置 Claude Code Router，自动生成配置文件和 alias 命令。</p>
-            <div class="feature-preview">
-                <div class="preview-item">✨ 可视化配置界面</div>
-                <div class="preview-item">🎯 智能推荐最佳配置</div>
-                <div class="preview-item">📋 一键生成所有命令</div>
-                <div class="preview-item">💾 配置文件导出功能</div>
-            </div>
-        </div>
-        
-        <div class="feature-card coming-soon">
-            <div class="coming-soon-badge">即将推出</div>
-            <h3>📊 性能对比分析</h3>
-            <p>详细的供应商性能对比，包括响应速度、准确性、成本效益等多维度分析。</p>
-            <div class="feature-preview">
-                <div class="preview-item">⚡ 响应速度测试</div>
-                <div class="preview-item">🎯 准确性评估</div>
-                <div class="preview-item">💰 成本效益分析</div>
-                <div class="preview-item">📈 实时性能监控</div>
-            </div>
-        </div>
-        
-        <div class="feature-card coming-soon">
-            <div class="coming-soon-badge">即将推出</div>
-            <h3>💰 价格信息中心</h3>
-            <p>实时更新的供应商价格信息，帮助您选择最经济实惠的 AI 服务。</p>
-            <div class="feature-preview">
-                <div class="preview-item">💵 实时价格更新</div>
-                <div class="preview-item">📊 成本计算器</div>
-                <div class="preview-item">🏷️ 优惠信息推送</div>
-                <div class="preview-item">📈 价格趋势分析</div>
-            </div>
-        </div>
-        
-        <div class="feature-card coming-soon">
-            <div class="coming-soon-badge">即将推出</div>
-            <h3>🔧 高级配置管理</h3>
-            <p>支持多环境配置管理、配置验证、自动切换等高级功能。</p>
-            <div class="feature-preview">
-                <div class="preview-item">🌍 多环境管理</div>
-                <div class="preview-item">✅ 配置验证</div>
-                <div class="preview-item">🔄 自动切换</div>
-                <div class="preview-item">📝 配置模板</div>
-            </div>
-        </div>
-    </div>
-    
-    <div class="feedback-section">
-        <h3>💡 功能建议</h3>
-        <p>如果您有其他功能建议或想法，欢迎通过 GitHub Issues 与我们分享！</p>
-        <a href="https://github.com/istarwyh/claude-code-router/issues" target="_blank" class="feedback-btn">
-            提交建议 →
-        </a>
-    </div>
-</div>
-`;
