+++ src/features/get-started/components/setup.ts
